pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chronoai"
version = "1.0.0"
description = "Personalized AI-powered desktop assistant for calendar reminders"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "ChronoAI Team"}
]
keywords = ["calendar", "assistant", "ai", "productivity", "desktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Desktop Environment",
    "Topic :: Office/Business :: Scheduling",
]

dependencies = [
    "PyQt6>=6.4.0",
    "apscheduler>=3.10.0",
    "google-api-python-client>=2.90.0",
    "google-auth>=2.16.0",
    "google-auth-oauthlib>=0.8.0",
    "requests>=2.28.0",
    "pyttsx3>=2.90",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/chronoai"
Repository = "https://github.com/yourusername/chronoai"
Documentation = "https://chronoai.readthedocs.io"
"Bug Tracker" = "https://github.com/yourusername/chronoai/issues"

[project.scripts]
chronoai = "chronoai.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/chronoai"]

[tool.hatch.build.targets.sdist]
include = ["src/chronoai"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=chronoai --cov-report=html --cov-report=term-missing"
